+++ rollup.config.js
@@ -33,7 +33,7 @@ export default {
       inlineWasm: process.env['SEPARATE_WASM'] !== 'true',
     }),
     typescript(),
-    commonjs({ include: ['src/*.ts', 'node_modules/**'] }),
+    commonjs({ include: ['src/*.ts', 'node_modules/**'],exclude:["taplo/**/*.ts"] }),
     resolve({ jsnext: true, preferBuiltins: true }),
   ],
 };
